+++ CHANGES.md
@@ -8,6 +8,9 @@
 
 ## Minor code helpers
 
+* `SessionListener` supports `sessionPeerIdentificationReceived` that is invoked once successful
+peer version data is received.
+
 ## Behavioral changes and enhancements
 
 * [SSHD-930](https://issues.apache.org/jira/browse/SSHD-930) - Added configuration allowing the user to specify whether client should wait